remove duplicate codes (#11588)
