+++ native/cocos/renderer/gfx-vulkan/VKGPUObjects.h
@@ -27,9 +27,9 @@
 
 #include "VKStd.h"
 #include "VKUtils.h"
-#include "base/CachedArray.h"
 #include "base/Log.h"
 #include "base/std/container/unordered_set.h"
+#include "core/memop/CachedArray.h"
 
 #define TBB_USE_EXCEPTIONS 0 // no-rtti for now
 #include "tbb/concurrent_unordered_map.h"
